refactor(server): move the middleware and startup helpers out of main.go
main.go held the HTTP middleware and the startup helpers alongside main
itself. Same package, pure moves with their doc comments:

- middleware.go: securityHeadersMiddleware, corsMiddleware,
  maxRequestSizeMiddleware, silentLogger, isLongRunningPath and
  streamingAwareTimeout (the package already had middleware_test.go and
  security_middleware_test.go)
- startup.go: initAppLogging, cleanupInterruptedRequests, warmCaches
  and initKeyCacheTTL

main.go goes from 860 to 479 lines and leaves the size allowlist. main
itself is untouched, so its funlen exclusion still reads 298 as the
config's comment says. Tests are unchanged.
