+++ cmd/server/main.go
@@ -3,17 +3,12 @@ package main
 // Package main is the entry point for the model-hotel LLM gateway server.
 
 import (
-	"bytes"
 	"context"
 	"embed"
-	"encoding/json"
-	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"os"
 	"os/signal"
-	"slices"
 	"strings"
 	"syscall"
 	"time"
@@ -21,24 +16,20 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5/pgxpool"
 
 	"github.com/hugalafutro/model-hotel/internal/admin"
 	"github.com/hugalafutro/model-hotel/internal/adminauth"
 	"github.com/hugalafutro/model-hotel/internal/alert"
 	"github.com/hugalafutro/model-hotel/internal/api"
 	"github.com/hugalafutro/model-hotel/internal/audit"
-	"github.com/hugalafutro/model-hotel/internal/auth"
 	"github.com/hugalafutro/model-hotel/internal/authcookie"
 	"github.com/hugalafutro/model-hotel/internal/clientip"
 	"github.com/hugalafutro/model-hotel/internal/config"
-	"github.com/hugalafutro/model-hotel/internal/ctxkeys"
 	"github.com/hugalafutro/model-hotel/internal/db"
 	"github.com/hugalafutro/model-hotel/internal/debuglog"
 	"github.com/hugalafutro/model-hotel/internal/events"
 	"github.com/hugalafutro/model-hotel/internal/failover"
 	"github.com/hugalafutro/model-hotel/internal/model"
-	"github.com/hugalafutro/model-hotel/internal/otelexport"
 	"github.com/hugalafutro/model-hotel/internal/provider"
 	"github.com/hugalafutro/model-hotel/internal/proxy"
 	"github.com/hugalafutro/model-hotel/internal/pwned"
@@ -56,56 +47,6 @@ var version = "dev"
 
 //go:embed all:static
 var staticFiles embed.FS
-
-// initAppLogging routes slog output through the app log pipeline so debuglog
-// calls reach the ring buffer and database (not just os.Stdout). When OTLP log
-// export is enabled (OTEL_EXPORTER_OTLP_ENDPOINT), fan out to it too so the
-// same structured records are pushed to an OpenTelemetry collector. Returns
-// the OTLP shutdown hook, nil when export is not enabled.
-func initAppLogging(ctx context.Context) func(context.Context) error {
-	appLogHandler := api.NewAppSlogHandler(debuglog.Level())
-	var otelLogShutdown func(context.Context) error
-	if otelexport.LogsEnabled() {
-		otelHandler, shutdown, oerr := otelexport.NewSlogHandler(ctx, "model-hotel", debuglog.Level())
-		if oerr != nil {
-			debuglog.Error("otel: OTLP log export init failed; continuing without it", "error", oerr)
-		} else {
-			appLogHandler = debuglog.NewFanout(appLogHandler, otelHandler)
-			otelLogShutdown = shutdown
-		}
-	}
-	debuglog.SetHandler(appLogHandler)
-	// Logged after SetHandler installs the fan-out, so the confirmation itself
-	// is also exported to the OTLP collector.
-	if otelLogShutdown != nil {
-		debuglog.Info("otel: OTLP log export enabled")
-	}
-	return otelLogShutdown
-}
-
-// cleanupInterruptedRequests marks request logs left in "pending" or
-// "streaming" state from a previous server crash, restart, or unhandled error
-// as failed. Using serverStartTime (captured before DB was ready) means we
-// only reclaim rows that predate this process — a genuine streaming request
-// that happens to be long-running is never touched.
-func cleanupInterruptedRequests(pool *pgxpool.Pool, serverStartTime time.Time) {
-	tag, err := pool.Exec(context.Background(), `
-		UPDATE request_logs
-		SET state = 'failed', error_kind = 'internal', error_message = 'request interrupted (server restart)'
-		WHERE state IN ('pending', 'streaming')
-		  AND created_at < $1`, serverStartTime)
-	if err == nil && tag.RowsAffected() > 0 {
-		debuglog.Info("startup: stale log cleanup", "rows", tag.RowsAffected())
-		events.Publish(events.Event{
-			Type:     "logs.stale_startup",
-			Severity: "warning",
-			Message:  fmt.Sprintf("Server restart interrupted %d pending %s", tag.RowsAffected(), util.Plural(int(tag.RowsAffected()), "request", "requests")),
-			Metadata: map[string]any{"count": tag.RowsAffected()},
-		})
-	} else if err != nil {
-		debuglog.Error("startup: stale log cleanup failed", "error", err)
-	}
-}
 
 func main() {
 	// Initialise the structured logger before anything can log, so the
@@ -534,327 +475,5 @@ func main() {
 		if err := otelLogShutdown(flushCtx); err != nil {
 			debuglog.Error("otel: OTLP log exporter shutdown failed", "error", err)
 		}
-	}
-}
-
-// securityHeadersMiddleware sets the standard security headers on every
-// response.
-func securityHeadersMiddleware(cfg *config.Config) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Header().Set("X-Content-Type-Options", "nosniff")
-			// When ALLOW_EMBED=true, X-Frame-Options and CSP frame-ancestors
-			// are omitted entirely so any origin can embed the page in an
-			// iframe (e.g. workspace browsers, Home Assistant).
-			if !cfg.AllowEmbed {
-				w.Header().Set("X-Frame-Options", "DENY")
-			}
-			w.Header().Set("Referrer-Policy", "strict-origin-when-cross-origin")
-			// HSTS only over TLS. Plain HTTP (e.g. behind a reverse proxy that
-			// terminates TLS) must not set HSTS or browsers will cache a broken
-			// redirect to a non-existent HTTPS listener. Currently the server
-			// only serves plain HTTP (ListenAndServe), so this guard is a
-			// forward-compatible placeholder: it will activate automatically if
-			// TLS is added later via ListenAndServeTLS.
-			if r.TLS != nil {
-				w.Header().Set("Strict-Transport-Security", "max-age=63072000; includeSubDomains; preload")
-			}
-			// CSP allows same-origin scripts/styles (needed for embedded SPA).
-			// Style 'unsafe-inline' is required for Vite's injected style tags (CSS-based
-			// animations and dynamic theme overrides). Script 'unsafe-inline' is NOT
-			// needed: Vite outputs module scripts, not inline ones.
-			if cfg.AllowEmbed {
-				w.Header().Set("Content-Security-Policy", "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data: blob:; connect-src 'self'; base-uri 'self'; form-action 'self'")
-			} else {
-				w.Header().Set("Content-Security-Policy", "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data: blob:; connect-src 'self'; frame-ancestors 'none'; base-uri 'self'; form-action 'self'")
-			}
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// corsMiddleware allows the configured origins (CORS_ORIGINS) and answers
-// preflight requests.
-func corsMiddleware(cfg *config.Config) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			origin := r.Header.Get("Origin")
-			if origin == "" {
-				next.ServeHTTP(w, r)
-				return
-			}
-
-			allowed := slices.Contains(cfg.CORSOrigins, origin)
-
-			w.Header().Set("Vary", "Origin")
-
-			if allowed {
-				w.Header().Set("Access-Control-Allow-Origin", origin)
-				w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-				w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-				w.Header().Set("Access-Control-Allow-Credentials", "true")
-				w.Header().Set("Access-Control-Max-Age", "86400")
-			}
-
-			if r.Method == "OPTIONS" {
-				w.WriteHeader(http.StatusNoContent)
-				return
-			}
-
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// maxRequestSizeMiddleware caps every request body at maxBytes.
-func maxRequestSizeMiddleware(maxBytes int64) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			r.Body = http.MaxBytesReader(w, r.Body, maxBytes)
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// warmCaches pre-warms caches synchronously before accepting connections.
-// Provider, model, and failover lookups are fast (simple SELECT queries),
-// but key warming (Argon2id) can take ~150ms per provider. The total
-// warm-up cost is typically under 1s for a handful of providers —
-// far better than letting the first request pay the cold-cache penalty
-// of ~170ms+ in failover + model + provider + key decryption DB queries.
-func warmCaches(deps discoveryDeps, settingsRepo *settings.Repository) {
-	ctx := context.Background()
-
-	providers, err := deps.providerRepo.List(ctx)
-	if err != nil {
-		debuglog.Error("cache: warm failed to list providers", "error", err)
-	} else {
-		enabledProviders := make([]*provider.Provider, 0, len(providers))
-		for _, p := range providers {
-			if !p.Enabled || !p.AutodiscoveryEnabled {
-				continue
-			}
-			if len(p.EncryptedKey) > 0 {
-				auth.WarmKeyCache(p.EncryptedKey, p.KeyNonce, p.KeySalt, deps.cfg.MasterKey)
-			}
-			enabledProviders = append(enabledProviders, p)
-		}
-		provider.WarmProviderCache(enabledProviders)
-	}
-
-	enabledModels, err := deps.modelRepo.ListEnabled(ctx)
-	if err != nil {
-		debuglog.Error("cache: warm failed to list models", "error", err)
-	} else {
-		model.WarmModelCache(enabledModels)
-	}
-
-	failoverGroups, err := deps.failoverRepo.List(ctx)
-	if err != nil {
-		debuglog.Error("cache: warm failed to list failover groups", "error", err)
-	} else {
-		failover.WarmFailoverCache(failoverGroups)
-	}
-
-	settingsRepo.WarmCache(ctx)
-
-	debuglog.Info("cache: key, provider, model, failover, and settings caches warmed")
-}
-
-// initKeyCacheTTL seeds the key cache TTL from settings and reacts to changes.
-func initKeyCacheTTL(settingsRepo *settings.Repository) {
-	auth.SetKeyCacheTTL(settingsRepo.GetDuration(context.Background(), "key_cache_ttl", auth.DefaultKeyCacheTTL))
-	settingsRepo.RegisterOnChange(func(key, value string) {
-		if key == "key_cache_ttl" {
-			d, err := time.ParseDuration(value)
-			if err != nil || d <= 0 {
-				debuglog.Warn("keycache: invalid key_cache_ttl setting, keeping current value", "value", value, "error", err)
-				return
-			}
-			auth.SetKeyCacheTTL(d)
-			debuglog.Info("keycache: TTL updated", "ttl", d)
-		}
-	})
-}
-
-// silentLogger is like chi's middleware.Logger but suppresses request log
-// lines for high-frequency polling endpoints that would flood docker logs.
-func silentLogger(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
-		t1 := time.Now()
-		next.ServeHTTP(ww, r)
-		duration := time.Since(t1)
-
-		path := r.URL.Path
-		isStatic := strings.HasPrefix(path, "/assets/") || strings.HasPrefix(path, "/favicon")
-		// Match the noise allowlist against a slash-normalized path so a trailing
-		// slash (from a client or a reverse proxy) can't defeat an exact match and
-		// leak the request back to Info. Root "/" is preserved.
-		np := path
-		if len(np) > 1 {
-			np = strings.TrimRight(np, "/")
-		}
-		isNoisy := np == "/health" ||
-			strings.HasPrefix(np, "/api/logs/app") ||
-			(np == "/api/logs" && r.Method == "GET") ||
-			(np == "/api/system" && r.Method == "GET") ||
-			(np == "/api/events" && r.Method == "GET") ||
-			(np == "/api/stats" && r.Method == "GET") ||
-			(np == "/api/stats/timeseries" && r.Method == "GET") ||
-			(np == "/api/stats/provider-distribution" && r.Method == "GET") ||
-			(np == "/api/models" && r.Method == "GET") ||
-			(np == "/api/providers" && r.Method == "GET") ||
-			// Fleet heartbeat: Front Desk pings every member ~every 2.5s with an
-			// announce POST and polls its version via GET /api/settings. Both are
-			// machine-to-machine liveness traffic, not human activity, and at
-			// ~24/min/member they otherwise flood app_logs (the App Logs page).
-			np == "/api/fleet/announce" ||
-			(np == "/api/settings" && r.Method == "GET")
-		if isStatic && ww.Status() < 400 {
-			return
-		}
-		// The path goes last in every branch below. It is caller-controlled, so
-		// a log reader scanning left to right meets every field the gateway
-		// vouches for before it reaches anything a visitor wrote. Values are
-		// escaped too; see quoteLogValue in internal/api/applogs_slog.go.
-		status := ww.Status()
-		switch {
-		case status >= 500:
-			debuglog.Error("access: request",
-				"method", r.Method,
-				"host", r.Host,
-				"remote", clientip.From(r),
-				"status", status,
-				"bytes", ww.BytesWritten(),
-				"duration", duration,
-				"path", r.URL.Path)
-		case status >= 400:
-			debuglog.Warn("access: request",
-				"method", r.Method,
-				"host", r.Host,
-				"remote", clientip.From(r),
-				"status", status,
-				"bytes", ww.BytesWritten(),
-				"duration", duration,
-				"path", r.URL.Path)
-		case isNoisy:
-			debuglog.Debug("access: request",
-				"method", r.Method,
-				"host", r.Host,
-				"remote", clientip.From(r),
-				"status", status,
-				"bytes", ww.BytesWritten(),
-				"duration", duration,
-				"path", r.URL.Path)
-		default:
-			debuglog.Info("access: request",
-				"method", r.Method,
-				"host", r.Host,
-				"remote", clientip.From(r),
-				"status", status,
-				"bytes", ww.BytesWritten(),
-				"duration", duration,
-				"path", r.URL.Path)
-		}
-	})
-}
-
-// streamingAwareTimeout returns middleware that sets a request deadline only
-// for non-streaming requests. Streaming LLM calls (e.g. code generation that
-// runs for 10+ minutes) must not be killed by a short server-side timeout.
-//
-// It works by peeking at the request body to check the "stream" field:
-//   - stream=true  → no context deadline (client disconnect detection still works)
-//   - stream=false/absent → context deadline of maxNonStreamingDur
-//
-// The request body is stored in the context so downstream handlers can
-// reuse it without a second allocation, and also restored as r.Body for
-// any handler that reads it directly.
-// isLongRunningPath reports whether the request targets a multimodal proxy
-// endpoint whose legitimate latency exceeds the non-streaming deadline:
-// image generation/edits and audio synthesis/transcription regularly take
-// minutes. The proxy's per-attempt failover timeout and overall deadline
-// still bound these requests.
-func isLongRunningPath(path string) bool {
-	return strings.HasPrefix(path, "/v1/images/") || strings.HasPrefix(path, "/v1/audio/")
-}
-
-func streamingAwareTimeout(maxNonStreamingDur time.Duration) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// Only POST /v1/chat/completions carries a stream flag;
-			// other routes (e.g. GET /v1/models) get the non-streaming timeout.
-			if r.Method != http.MethodPost {
-				ctx, cancel := context.WithTimeout(r.Context(), maxNonStreamingDur)
-				defer cancel()
-				next.ServeHTTP(w, r.WithContext(ctx))
-				return
-			}
-
-			// Multipart uploads (audio transcription/translation, image
-			// edits/variations) are never buffered here: reading megabytes
-			// into memory before auth would let unauthenticated clients pin
-			// large allocations, and the JSON peek cannot apply anyway (the
-			// model field lives in the form, parsed by the handler after
-			// auth). These routes are long-running, so no deadline either.
-			if strings.HasPrefix(strings.ToLower(r.Header.Get("Content-Type")), "multipart/") {
-				if isLongRunningPath(r.URL.Path) {
-					next.ServeHTTP(w, r)
-					return
-				}
-				ctx, cancel := context.WithTimeout(r.Context(), maxNonStreamingDur)
-				defer cancel()
-				next.ServeHTTP(w, r.WithContext(ctx))
-				return
-			}
-
-			parseStart := time.Now()
-			body, err := io.ReadAll(r.Body)
-			_ = r.Body.Close()
-			if err != nil {
-				util.WriteOpenAIError(w, "failed to read request body", http.StatusBadRequest)
-				return
-			}
-
-			// Extract both stream and model in a single unmarshal so
-			// downstream handlers can skip re-parsing cached bytes. The peek
-			// runs regardless of Content-Type: clients send JSON chat bodies
-			// with text/plain or form-urlencoded headers, and skipping them
-			// would wrongly impose the non-streaming deadline on their streams.
-			var parsed struct {
-				Stream bool   `json:"stream"`
-				Model  string `json:"model"`
-			}
-			isStreaming := false
-			modelName := ""
-			if json.Unmarshal(body, &parsed) == nil {
-				isStreaming = parsed.Stream
-				modelName = parsed.Model
-			}
-			parseMs := float64(time.Since(parseStart).Microseconds()) / 1000.0
-
-			// Restore the body so downstream handlers can read it
-			r.Body = io.NopCloser(bytes.NewReader(body))
-
-			// Store body bytes + extracted fields + timing in context
-			ctx := context.WithValue(r.Context(), ctxkeys.RequestBodyKey, body)
-			ctx = context.WithValue(ctx, ctxkeys.RequestBodyParseMsKey, parseMs)
-			ctx = context.WithValue(ctx, ctxkeys.RequestModelKey, modelName)
-			ctx = context.WithValue(ctx, ctxkeys.IsStreamingKey, isStreaming)
-
-			// Long-running multimodal routes (image generation, audio) get the
-			// streaming treatment even without a body stream flag: their
-			// legitimate latencies (image models, large transcriptions, SSE
-			// synthesis) exceed the non-streaming deadline. The proxy's
-			// per-attempt failover timeout still bounds each upstream call.
-			if isStreaming || isLongRunningPath(r.URL.Path) {
-				next.ServeHTTP(w, r.WithContext(ctx))
-			} else {
-				ctx, cancel := context.WithTimeout(ctx, maxNonStreamingDur)
-				defer cancel()
-				next.ServeHTTP(w, r.WithContext(ctx))
-			}
-		})
 	}
 }
