+++ scripts/ci/size-allowlist.txt
@@ -7,7 +7,6 @@
 # over the ceiling gets split instead.
 #
 # Format: path<TAB>lines
-cmd/server/main.go	860
 internal/api/applogs_test.go	2083
 internal/api/discovery.go	885
 internal/api/discovery_integration_quota_test.go	2735
